--adc for battery ok

+++ user/driver/timer.h
@@ -4,6 +4,6 @@
 
 void TIM3_Int_Init(void);
 void TIM2_Int_Init(void);
-void delay_us(uint32_t us_cnt); 
-void delay_ms(uint32_t ms_cnt);
+void delay_us(uint16_t us_cnt); 
+void delay_ms(uint16_t ms_cnt);
 #endif
